requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# AI and NLP
openai==1.3.7
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
keybert==0.7.0
sentence-transformers==2.2.2

# Document processing
pypdf==3.17.1
python-docx==1.1.0
reportlab==4.0.7
Pillow==10.1.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2

# Configuration and utilities
pyyaml==6.0.1
jinja2==3.1.2
python-dateutil==2.8.2

# HTTP and networking
httpx==0.25.2
aiofiles==23.2.1

# Optional: For enhanced PDF processing
pdfplumber==0.10.3

# Optional: For enhanced text processing
transformers==4.36.0
torch==2.1.1

# Optional: For web scraping job descriptions
beautifulsoup4==4.12.2
requests==2.31.0 